src/requirements.txt
fastapi[standard]
pydantic-settings
email-validator
pyyaml
sqlalchemy
alembic
psycopg2-binary
passlib[bcrypt]
flask
flask-sqlalchemy
Flask-WTF
wtforms[email]
jinjax
basic-components[utils]
Flask-WTF
wtforms[email]
requests
resend
cryptography
pyJWT
python-multipart